requirements.txt
python-crontab~=3.0.0
croniter~=2.0.1
fastapi~=0.109.0
uvicorn~=0.25.0
aiogram~=3.3.0
psycopg2~=2.9.9